requirements.txt
streamlit>=1.30
requests>=2.31
pandas>=2.0
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
pydantic>=2.5
python-dotenv>=1.0